+++ _config.php
@@ -1,12 +1,9 @@
 <?php
 /**
- * ColorPicker Form-Field for SilverStripe 3.x
- * @author Roman Schmid, AKA bummzack
- * 
  * ColorPicker jQuery widget/plugin by http://www.eyecon.ro/colorpicker/
  */
 
 // define colorpicker dir so that the installation doesn't need a fixed directory name
-if(!defined('COLORPICKER_DIR')){
-	define('COLORPICKER_DIR', basename(__DIR__));
+if (!defined('COLORPICKER_DIR')) {
+    define('COLORPICKER_DIR', basename(__DIR__));
 }