+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # silverstripe-colorpicker
 
+## 4.0
+
+ - SilverStripe 4.x compatible
+
+## 3.0
+
+ - SilverStripe 3.x compatible.
+ - Added custom field-type `Color`
+
 ## 1.0
 
 Initial import from http://bummzack.ch/colorpicker/